Move ProofGenForm to own component directory
